+++ examples/dagster_examples/intro_tutorial/advanced/scheduling/scheduler.py
@@ -1,12 +1,6 @@
 import csv
 
-from dagster import (
-    RepositoryDefinition,
-    ScheduleDefinition,
-    pipeline,
-    schedules,
-    solid,
-)
+from dagster import RepositoryDefinition, ScheduleDefinition, pipeline, solid
 from dagster.utils import file_relative_path
 
 
@@ -33,7 +27,6 @@ def cereal_repository():
     )
 
 
-@schedules
 def cereal_schedules():
     return [
         ScheduleDefinition(